Refactored code to include a navigation bar that is seen on each page, whose links are tailored to what the user is trying to accomplish. Added basic CSS styling.
