+++ front-end/src/App.css
@@ -1,40 +1,17 @@
-.App {
-  text-align: center;
-  height: 100%;
+ul {
+list-style-type: none;
+display: flex;
+justify-content: space-between;
+align-content: center;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+#right-nav {
+display: flex;
+flex-flow: row wrap;
+justify-content: flex-end;
+margin-right: 1rem;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
+#right-nav li {
+margin: auto 1em;
 }
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
-}
-
