+++ front-end/src/components/layout/LoginPage.js
@@ -2,6 +2,7 @@ import React, { useState } from "react";
 import Form from 'react-bootstrap/Form';
 import Button from 'react-bootstrap/Button';
 
+
 export default function Login(){
 
     const [email, setEmail] = useState("");
